// src/theme/types.ts
export interface DatePickerTheme {
  fontFamily?: string
  fontSize?: string
  primaryColor?: string
  primaryTextColor?: string
  rangeColor?: string
  textColor?: string
  mutedTextColor?: string
  backgroundColor?: string
  surfaceColor?: string
  borderColor?: string
  borderRadius?: string
  daySize?: number
  shadow?: string
}